requirements.txt
streamlit>=1.31.0
tensorflow-cpu>=2.16.1
numpy<2.0.0
pillow